Make TollCalculator configurable
Adds TollCalculatorConfig to be injected into TollCalculator. Config contains values that are most likely to change, like toll, schedule and exempt vehicle types. It notably contains a collection of rule sets. Currently, the last rule set found that is after the passage date is used. This is to make it possible to change the values of the business logic for all passages, without necessarily applying them if the code is called for older date that uses the old values.